Grasswalker/requirements.txt
alabaster==0.7.6
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
awscli==1.10.58
Babel==2.0
bleach==2.1.1
boto3==1.4.0
botocore==1.4.48
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.6
colorama==0.3.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==1.9
cycler==0.10.0
Cython==0.26.1
decorator==4.1.2
defusedxml==0.5.0
dill==0.2.7.1
Django==2.0.6
django-allauth==0.36.0
django-cors-headers==2.2.1
django-filter==1.1.0
django-pipeline==1.5.4
django-publications==0.6.1
django-rest-auth==0.9.3
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
docutils==0.12
entrypoints==0.2.3
eventlet==0.19.0
feedparser==5.2.1
Flask==0.11.1
Flask-WTF==0.13.1
future==0.16.0
futures==3.0.3
greenlet==0.4.10
html5lib==1.0b10
httpie==0.9.9
hyperopt==0.1
idna==2.7
inflection==0.3.1
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.1
itsdangerous==0.24
itypes==1.1.0
jedi==0.11.0
Jinja2==2.8
jmespath==0.9.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jwt-auth==0.4.3
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.0.0
mistune==0.7.4
mod-wsgi==4.6.4
more-itertools==3.2.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.1
nose==1.3.7
notebook==5.1.0
numpy==1.12.1
oauthlib==2.1.0
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.0.0
pillowfight==0.2
prompt-toolkit==1.0.15
protobuf==3.2.0
ptyprocess==0.5.2
pyasn1==0.1.9
pycparser==2.18
Pygments==2.2.0
pyhocon==0.3.35
PyJWT==1.6.4
pymongo==3.6.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-engineio==1.0.3
python-socketio==1.6.0
python3-openid==3.1.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.1
Quandl==3.2.0
requests==2.19.1
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.1.2
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
snap==0.5
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
tensorflow==1.0.1
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
UNKNOWN==0.0.0
uritemplate==3.0.0
urllib3==1.23
virtualenv==15.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.11.11
widgetsnbextension==3.0.3
WTForms==2.1
